Removing static hostname
Find the hostname string that we should use for the zabbix sender from
the local *nix hostname call. Remove the .local string from the end if
it exists. This matches how we name out hosts on Zabbix. Might be worth
adding this as attribute to the command at some point, but we don't need
it right now.